Tenant database switch now supports Laravel Octane
