+++ src/Multitenancy.php
@@ -1,0 +1,94 @@
+<?php
+
+namespace Spatie\Multitenancy;
+
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Support\Facades\Event;
+use Laravel\Octane\Events\RequestReceived as OctaneRequestReceived;
+use Laravel\Octane\Events\RequestTerminated as OctaneRequestTerminated;
+use Laravel\Octane\Octane;
+use Spatie\Multitenancy\Actions\MakeQueueTenantAwareAction;
+use Spatie\Multitenancy\Concerns\UsesMultitenancyConfig;
+use Spatie\Multitenancy\Models\Concerns\UsesTenantModel;
+use Spatie\Multitenancy\Models\Tenant;
+use Spatie\Multitenancy\Tasks\TasksCollection;
+use Spatie\Multitenancy\TenantFinder\TenantFinder;
+
+class Multitenancy
+{
+    use UsesTenantModel,
+        UsesMultitenancyConfig;
+
+    public function __construct(public Application $app)
+    {
+    }
+
+    public function start(): void
+    {
+        $this
+            ->registerTenantFinder()
+            ->registerTasksCollection()
+            ->configureRequests()
+            ->configureQueue();
+    }
+
+    public function end(): void
+    {
+        Tenant::forgetCurrent();
+    }
+
+    protected function determineCurrentTenant(): void
+    {
+        if (! $this->app['config']->get('multitenancy.tenant_finder')) {
+            return;
+        }
+
+        /** @var \Spatie\Multitenancy\TenantFinder\TenantFinder $tenantFinder */
+        $tenantFinder = $this->app[TenantFinder::class];
+
+        $tenant = $tenantFinder->findForRequest($this->app['request']);
+
+        $tenant?->makeCurrent();
+    }
+
+    protected function registerTasksCollection(): self
+    {
+        $this->app->singleton(TasksCollection::class, function () {
+            $taskClassNames = $this->app['config']->get('multitenancy.switch_tenant_tasks');
+
+            return new TasksCollection($taskClassNames);
+        });
+
+        return $this;
+    }
+
+    protected function registerTenantFinder(): self
+    {
+        if ($this->app['config']->get('multitenancy.tenant_finder')) {
+            $this->app->bind(TenantFinder::class, $this->app['config']->get('multitenancy.tenant_finder'));
+        }
+
+        return $this;
+    }
+
+    protected function configureRequests(): self
+    {
+        if (! $this->app->runningInConsole()) {
+            $this->determineCurrentTenant();
+        }
+
+        return $this;
+    }
+
+    protected function configureQueue(): self
+    {
+        $this
+            ->getMultitenancyActionClass(
+                actionName: 'make_queue_tenant_aware_action',
+                actionClass: MakeQueueTenantAwareAction::class
+            )
+            ->execute();
+
+        return $this;
+    }
+}
